New Weapon System
New Weapon System
New Melee System
Started Making WinScreen

+++ Assets/Scripts/Player/Health/DamageOnCollision.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class DamageOnCollision : MonoBehaviour
 {
     [Header("Settings")]
-    [SerializeField] private float damage;
+    [SerializeField] private int damage;
 
     private void OnCollisionEnter(Collision other)
     {
